+++ api-demo/assets/advanced-input.ts
@@ -1,0 +1,21 @@
+export class Statistics {
+  /**
+   * Returns the average of two numbers.
+   * This incomplete HTML tag should be reported as an error: <tag
+   *
+   * @remarks
+   * Here's a {@customInline some data}.
+   *
+   * @customBlock
+   * Here's a custom block tag.
+   *
+   * @param x - The first input number
+   * @param y - The second input number
+   * @returns The average of `x` and `y`
+   *
+   * @beta @customModifier
+   */
+  public static add(x: number, y: number): number {
+    return x + y;
+  }
+}