+++ api-demo/assets/simple-input.ts
@@ -1,0 +1,17 @@
+
+  /**
+   * Returns the average of two numbers.
+   *
+   * @remarks
+   * This method is part of the {@link core-library/Statistics | Statistics subsystem}.
+   * This incomplete HTML tag should be reported as an error: <tag
+   *
+   * @privateRemarks
+   * This content should @not show up on the web site.
+   *
+   * @param x - The first input number
+   * @param y - The second input number
+   * @returns The average of `x` and `y`
+   *
+   * @beta
+   */